\name{tSVE-package}
\alias{tSVE-package}

\docType{package}
\title{
    \packageTitle{tSVE}
}
\description{
    \packageDescription{tSVE}
}
\details{

\packageDESCRIPTION{tSVE}
    \packageIndices{tSVE}
}
\author{
    \packageAuthor{tSVE}

    Maintainer: \packageMaintainer{tSVE}
}

\keyword{ package }

