\name{TVTB-package}
\alias{TVTB-package}

\docType{package}
\title{
    \packageTitle{TVTB}
}
\description{
    \packageDescription{TVTB}
}
\details{

\packageDESCRIPTION{TVTB}
    \packageIndices{TVTB}
}
\author{
    \packageAuthor{TVTB}

    Maintainer: \packageMaintainer{TVTB}
}

\keyword{ package }

